Bug 1501005 - refactor common code into a base class
